# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for cloud-init clients using the OpenCHAMI cloud-init server",
        "title": "OpenCHAMI Cloud-Init Server API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/OpenCHAMI/.github/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/admin/cluster-defaults": {
            "get": {
                "description": "Get default meta-data values for cluster.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "cluster-defaults"
                ],
                "summary": "Get cluster defaults",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cistore.ClusterDefaults"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Set default meta-data values for cluster.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "cluster-defaults"
                ],
                "summary": "Set cluster defaults",
                "parameters": [
                    {
                        "description": "Cluster defaults data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cistore.ClusterDefaults"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/admin/impersonation/{id}/meta-data": {
            "get": {
                "description": "Get meta-data for requesting node based on the requesting IP.\n\nIf the impersonation API is enabled, an ID can be provided in\nthe URL path using `/admin/impersonation`. In this case, the\nmeta-data will be retrieved for the requested ID.",
                "produces": [
                    "application/x-yaml"
                ],
                "summary": "Get meta-data for requesting node",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.MetaData"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/admin/instance-info/{id}": {
            "put": {
                "description": "Set meta-data for a specific node ID, overwriting relevant group meta-data.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "instance-data"
                ],
                "summary": "Set node-specific meta-data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Instance info data",
                        "name": "instance-info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cistore.OpenCHAMIInstanceInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/meta-data": {
            "get": {
                "description": "Get meta-data for requesting node based on the requesting IP.\n\nIf the impersonation API is enabled, an ID can be provided in\nthe URL path using `/admin/impersonation`. In this case, the\nmeta-data will be retrieved for the requested ID.",
                "produces": [
                    "application/x-yaml"
                ],
                "summary": "Get meta-data for requesting node",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.MetaData"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/openapi.json": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Return JSON-formatted OpenAPI documentation",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/phone-home/{id}": {
            "post": {
                "description": "Signal to the cloud-init server that the specific host has completed running\nthe cloud-init configuration tasks so that, if a WireGuard tunnel is being used,\nit can be torn down. This endpoint should not be manually requested by a user\nbut is only meant to be used by a cloud-init client that has received its\nconfig from an OpenCHAMI cloud-init server.",
                "tags": [
                    "phone-home"
                ],
                "summary": "Signal to cloud-init server that host has completed running cloud-init configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node's unique identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Node's WireGuard RSA public key",
                        "name": "pub_key_rsa",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Node's WireGuard ECDSA public key",
                        "name": "pub_key_ecdsa",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Node's WireGuard ED35519 public key",
                        "name": "pub_key_ed25519",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Node's given instance ID",
                        "name": "instance_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Node's given hostname",
                        "name": "hostname",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Node's given fully-qualified domain name",
                        "name": "fqdn",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "cistore.ClusterDefaults": {
            "type": "object",
            "properties": {
                "availability-zone": {
                    "type": "string"
                },
                "base-url": {
                    "type": "string",
                    "example": "http://demo.openchami.cluster:8081/cloud-init"
                },
                "boot-subnet": {
                    "type": "string"
                },
                "cloud_provider": {
                    "type": "string"
                },
                "cluster-name": {
                    "type": "string",
                    "example": "demo"
                },
                "nid-length": {
                    "type": "integer",
                    "example": 3
                },
                "public-keys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIMLtQNuzGcMDatF+YVMMkuxbX2c5v2OxWftBhEVfFb+U user1@demo-head",
                        "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIB4vVRvkzmGE5PyWX2fuzJEgEfET4PRLHXCnD1uFZ8ZL user2@demo-head"
                    ]
                },
                "region": {
                    "type": "string"
                },
                "short-name": {
                    "type": "string",
                    "example": "nid"
                },
                "wg-subnet": {
                    "type": "string"
                }
            }
        },
...
        "cistore.OpenCHAMIInstanceInfo": {
            "type": "object",
            "properties": {
                "availability-zone": {
                    "type": "string"
                },
                "cloud-init-base-url": {
                    "type": "string"
                },
                "cloud-provider": {
                    "type": "string"
                },
                "cluster-name": {
                    "type": "string",
                    "example": "demo"
                },
                "hostname": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "x3000c1b1n1"
                },
                "instance-id": {
                    "type": "string"
                },
                "instance-type": {
                    "type": "string"
                },
                "local-hostname": {
                    "type": "string",
                    "example": "compute-1"
                },
                "public-keys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIMLtQNuzGcMDatF+YVMMkuxbX2c5v2OxWftBhEVfFb+U user1@demo-head",
                        "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIB4vVRvkzmGE5PyWX2fuzJEgEfET4PRLHXCnD1uFZ8ZL user2@demo-head"
                    ]
                },
                "region": {
                    "type": "string"
                }
            }
        },
        "main.Group": {
            "type": "object",
            "additionalProperties": true
        },
        "main.InstanceData": {
            "type": "object",
            "properties": {
                "v1": {
                    "type": "object",
                    "properties": {
                        "availability-zone": {
                            "type": "string"
                        },
                        "cloud-name": {
                            "type": "string"
                        },
                        "cloud-provider": {
                            "type": "string"
                        },
                        "hostname": {
                            "type": "string"
                        },
                        "instance-id": {
                            "type": "string"
                        },
                        "instance-type": {
                            "type": "string"
                        },
                        "local-hostname": {
                            "type": "string"
                        },
                        "local-ipv4": {},
                        "public-keys": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "region": {
                            "type": "string"
                        },
                        "vendor-data": {
                            "$ref": "#/definitions/main.VendorData"
                        }
                    }
                }
            }
        },
        "main.MetaData": {
            "type": "object",
            "properties": {
                "cluster-name": {
                    "type": "string",
                    "example": "demo"
                },
                "hostname": {
                    "type": "string",
                    "example": "compute-1.demo.openchami.cluster"
                },
                "instance-data": {
                    "$ref": "#/definitions/main.InstanceData"
                },
                "instance-id": {
                    "type": "string"
                },
                "local-hostname": {
                    "type": "string",
                    "example": "compute-1"
                }
            }
        },
        "main.VendorData": {
            "type": "object",
            "properties": {
                "cabinet": {
                    "type": "string"
                },
                "cloud-init-base-url": {
                    "type": "string"
                },
                "cluster_name": {
                    "type": "string",
                    "example": "demo"
                },
                "groups": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/main.Group"
                    }
                },
                "location": {
                    "type": "string"
                },
                "nid": {
                    "type": "integer"
                },
                "rack": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "sub-role": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
    }
}